v1.0.0 - correção de bug do modal editar
